extract parseIncomingMessage out of NodeExecutionFunctions
